Added Convert Checklists to Cards
Currently is available on every card. Requires that the Checklist share
names with target list. Creates creates cards with name based on the
list item and also generates a description of what its parent card was
as well as that cards list. Story cards will include the story id in the
name automatically.